pyproject.toml
[tool.poetry]
name = "access-token"
version = "0.1.0"
description = ""
authors = ["Phong Do <haiphong3107@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.104.1,<0.105.0"
pyyaml = ">=6.0.1,<6.1.0"
sqlalchemy = ">=2.0.23,<2.1.0"
psycopg = {version = ">=3.1.12,<3.2.0", extras = ["binary"]}
uvicorn = {version = ">=0.22.0,<0.23.0", extras = ["standard"]}
starlette = ">=0.27.0,<0.28.0"
msal = ">=1.23.0,<1.24.0"
argon2-cffi = ">=23.1.0,<23.2.0"
jinja2 = ">=3.1.2,<3.2.0"
pytest = ">=7.4.3,<7.5.0"
pytest-postgresql = ">=5.0.0,<5.1.0"
pytest-mock = ">=3.12.0,<3.13.0"
mock = ">=5.1.0,<5.2.0"
cryptography = ">=41.0.4,<41.1.0"
requests = ">=2.28.1,<2.29.0"
schedule = ">=1.2.1,<1.3.0"
httpx = ">=0.27.0,<0.28.0"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
